Global Theme Colors
Add the ability to quickly change the hue of all the colors in the theme. Nothing has been changed behind the scenes and every color can still be individually tweaked. (Additionally colors can be locked from being globally modified)

Added a function to apply the theme in a particular theme manager from code.