Not Optimization
Added a function called commit which optimizes a copy of the parse tree (and will eventually save the result in the list).

Added the first optimization pass (elimination of NOTs). This is done to AND and OR nodes by applying De'Morgans, and to single Literal nodes by creating an OR node with the other two Literal options.